env flex, lightening os lib, env checking, loadenv

+++ getter.py
@@ -1,4 +1,5 @@
-import subprocess, requests, time, sys, os
+import subprocess, requests, time, sys 
+from os import getenv, system, path
 
 def load_env():
     from dotenv import load_dotenv
@@ -6,11 +7,14 @@ def load_env():
     try:
         with open(f'/boot/{env}', 'r') as file:
             file.read()
-
-        return load_dotenv(f'/boot/{env}')
+            file.close()
+            env = f'/boot/.env'
 
     except:
         return load_dotenv(env)
+
+def envCheck(env_file):
+    return path.exists(env_file)
 
 def getCloudVersion(host, header):
     version = requests.get(host, headers=header)
@@ -36,8 +40,10 @@ def main():
     return sketch
 
 if __name__ == "__main__":
+    from dotenv import load_dotenv, find_dotenv
+
     try:
-        # cloud_version = getCloudVersion(os.getenv('CLOUD_HOST'), {'x-siap-token': 'xxx'})
+        # cloud_version = getCloudVersion(getenv('CLOUD_HOST'), {'x-siap-token': 'xxx'})
         # print(f'[VERSION] cloud version: {cloud_version}')
 
         # local_version = getLocalVersion('.ver', 'r')
@@ -48,15 +54,17 @@ if __name__ == "__main__":
         #     time.sleep(1)
         #     up = getLocalVersion('.ver', 'w')
 
+        if envCheck('/media/bca_env/env.txt'): system('cp /media/bca_env/env.txt .env')
+
+        if not envCheck('.env'): raise RuntimeError("[ERROR] Env not found on any local system!")
+
         load_env()
         exec(main())
         gateway.makeLog(2000, 'reboot')
-        
-        # threading.Thread(target=camera, args=(), daemon=True).start()
         loop()
 
     except RuntimeError as r:
-        sys.exit(r)
+        sys.exit(r) 
 
     except Exception as e:
         gateway.reboot(io.close(), e)
